Support eSpace pool participate conflux governance (#79)
* reorg file

* check info

* refactor code

* add espace voting escrow-related methods

* rewrite deploy scripts

* update readme

* complete readme

* update docs

* optimize the repo readme

* refactor contract deploy doc

* update contract upgrade doc

+++ contract/contracts/PoolManager.sol
@@ -29,7 +29,7 @@ contract PoolManager is Ownable {
       IPoSPool poolInstance = IPoSPool(poolAddresses[i]);
       poolInfos[i] = PoolInfo({
         poolAddress: poolAddresses[i],
-        totalAvailable: poolInstance.poolSummary().available,
+        totalAvailable: uint64(poolInstance.poolSummary().available),
         name: poolInstance.poolName(),
         apy: poolInstance.poolAPY(),
         fee: poolInstance.poolUserShareRatio()
